Prefer "user-breakpoint" over "explicit" when talking about breakpoints
Summary:
# Context
We set breakpoints on the VM for one of two reasons: because a user requested it, or because they are supporting stepping of a process. Internally, we call the first type "explicit" breakpoints and use "step" breakpoints for the others.

# Problem
The name "explicit" can be a bit confusing. In particular, as we plan to add to support function-breakpoints, we'd like to consider them also "explicit", but the name somehow doesn't feel right.

# This diff
We now use `user-breakpoint` instead of explicit. This is a mechanical change, no behaviour changes

Reviewed By: TD5

Differential Revision: D88070258

fbshipit-source-id: 0dcd15e1cb6098bc0704aaec7b8820d8a207d67d