Simplify internal representation of breakpoints_hit
Summary:
# Context
In `edb_server_break` we handle all the book-keeping about breakpoints. In particular, we keep a set of all processes currently stopped on a breakpoint. This is exposed in the `edb` API is mostly useful for testing or working from the shell

# Problem
In the stored set of breakpoints-hit, we are currently using the "vm-module", that is, the actual module in the VM where the breakpoint happened, tha,t, in the presence of mocking, may not be the module the user care about.

This means that we need to make conversions every time we return this information to the user, and that we need a duplicate type definition to account for the version with `vm_module()` and with `module()`.

# This diff
Simplify things by storing the `module()` relevant to the user in the breakpoints-hit set, so no further conversions are necessary

Reviewed By: michalmuskala

Differential Revision: D88139896

fbshipit-source-id: a04b58334c7b4f63a82e0073e01f23f00445973d