Prepare internal breakpoints api for function breakpoints
Summary:
# Context
At the moment, the user can only set breakpoints on lines, but we would like to extend it so that they can set breakpoints on functions as well.

The internal `edb_server_break` module handles all the book-keeping about what breakpoints in the VM were set for what.

# Problem
This module It has functions to query all breakpoints set by the user, but at the moment these return a set of lines. We'd like this functions to work the same way once we add function breakpoints.

# This diff
We change their return type so that they return a `breakpoints_info()` instead of a line.

Reviewed By: michalmuskala

Differential Revision: D88142622

fbshipit-source-id: 03fa3b857e3497d0466498c119526d632a4640a5