Prepare breakpoint-info api for function breakpoints
Summary:
# Context
So far the user can only set breakpoints on lines, but we want to extend the API so that we can set breakpoints on functions as well.

# Problem
We currently have functions to return all the breakpoints set or currently hit and they only return line information. We would like to reuse this api and be able to get information about function-breakpoints set or hit.

# This diff
We extend the result type of these operations with ` type` field and set it to `line` for now. Later we will be able to return breakpoint info of a different type.

Reviewed By: michalmuskala

Differential Revision: D88073405

fbshipit-source-id: f1d3252cecc3b862e45c7f19c5e018eeec5fc0a1