Fix breakpoint cleanup logic on termination
Summary:
# Context
The `edb_server` is a gen-server that acts as debugger process in the debuggee node. When it terminates, as a cleanup, it tries to unset an breakpoint set in the VM.

# Problem
For historical reasons, this clean-up was only removing breakpoints set by the user, but not those that were set for the sake of implementing a stepping operation.

Lingering breakpoints are not that problematic in practice, as in the absence of a debugger process the VM ignores them, but they could confuse a debugger process that later attaches.

# This diff
Change the logic so that now we remove every breakpoint

Reviewed By: TD5

Differential Revision: D88144173

fbshipit-source-id: f71bfad112a8035544c8a2164605d465ad351a95