Remove references to vm_module() from edb_server_break API
Summary:
# Context
The support for mocking is based on the idea of "substitute modules", e.g., when the user talks about module `foo`, internally we map it to the substitute `foo_meck` module.

To try to to avoid errors, we rely on the type-system to distinguish between a user provided module, a `module()`, and a module that corresponds directly to a module in the VM, `vm_module() :: {vm_module, module()}`. Because of the tag, they can't be mixed

# Problem
The idea was that the `vm_module()` are mostly an implementation detail inside the `edb_server_break` module, something we use to avoid errors. In practice, it ended up leaking the API completely: every function of the module receives and returns things of type `vm_module()` and the caller is responsible for converting back and forth, using conversion functions coming from the `edb_server_break` itself.

# This diff
Clean things up and make it so that, outside `edb_server_break` `vm_module()` doesn't exist. Well, almost, we still need to map modules in stack-frames back to user modules, so for that use-case, we still need to expose a `from_vm_module/2` function.

Reviewed By: TD5

Differential Revision: D88090708

fbshipit-source-id: e0b1e6fa55225b785ac3a273382b17b54972d0d4