Avoid possible crash when substitutting modules (mock support)
Summary:
# Context
The way we support mocking techniches like `meck` is by having an internal notion of module substitutes: whenever the user refers to module `foo`, internally we act on the substitute module, say, `foo_meck`, etc.

The API we expose for setting substitute modules tries to reapply any existing breakpoints in the original module to the substitute

# Problem
If applying reapplying breakpoints fail for any reason (e.g. substitute not compiled with debugging support), then the operation will crash, which can be confusing to the user

# This diff
Instead, we now proceed with the substitute.

Why not return an error instead? While we could, ithe overal behaviour would be inconsistent. In particular, it would mean that, if the substitute module does not support debugging, we'll fail only when breakpoints are initially set in the original module, but otherwise leave the user with a substitute that can't be debugged.

For now it is easier to just let it pass than try to detect that the module is not debuggable.

Reviewed By: robertoaloi

Differential Revision: D88090707

fbshipit-source-id: 9725c21ece7783176d1b4824d32915cdee40fa43