Prepare edb_server_break api for function breakpoints
Summary:
# Context
We would like to add support for "function breakpoints". The way this will work, for each "function breakpoint" we'll set one or more line-breakpoints, one per breakpoint. This means that we will need to adjust the book-keeping of breakpoints in `edb_server_break` to account for this.

# Problem
At the moment we have functions in `edb_server_break` to add/remove user breakpoints, and we'd like to re-use them to add either line or function breakpoints, to keep the API small. However, wit the current arity, we can't reuse them

# This diff
Make it so that the functions to add breakpoints in `edb_server_break` take module-line as a tuple. This will later allow us to extend them so they can receive either such a pair or an MFA

Reviewed By: michalmuskala

Differential Revision: D88070255

fbshipit-source-id: c4ba95280e35496b1f19132e81e3d04dd2325f50