Use consistent variable naming in edb_server_break
Summary:
In `edb_server_break` we keep track of all the breakpoints set in the VM, as we we can set them because the user asked for them, or for stepping.

The module also keeps track of "substitute" modules, that are used to support mocking, e.g. `foo_meck` is a substitute module for `meck`.

To avoid confusing modules and substitute modules, we have a tagged-type `vm_module()` and ensure that breakpointing and book-keeping is done in terms of `vm_module()`, and explicit conversions are needed from user `module()` to potentially substituted `vm_module()`.

We tend to use `Module` in variable names for things of type `module()` and `VmModule` for things of type `vm_module()`. But for historical reasons, we are not consistent, and there are still many functions that take about a `Module` when they actually work on a `vm_module()`.

This makes the code confusing and harder to understand than it should

Find all places where we are using `Module` for a `vm_module()` and replace it by `VmModule`.

This diff only changes variable names, there's no functionality change

Reviewed By: michalmuskala

Differential Revision: D88268562

fbshipit-source-id: 1ce77ff0e489aa499ac4fc62b24d18a2a6776698