Add set_function_breakpoits function
Summary:
# Context
In this stack we have added functions to the `edb` module to add and clear a function breakpoint. We now want to add the DAP support for this.

# Problem
Just like for line-breakpoints, in DAP there is a request to set all function breakpoints in one go, clearing any previously existing function-breakpoint as well. We don't have a way to implement this operation "atomically" the in the edb_server

# This diff
Just like for line-brakpoints we have a `edb:set_breakpoints()` function that does this, we introduce a `edb:set_function_breakpoints()`. In the `edb_server_break` module that handles the book-keeping of breakpoints, we need to add a function to clear all user-brakpoints across all modules to support this.

Reviewed By: TheGeorge

Differential Revision: D88588037

fbshipit-source-id: 140d390483ff88367cdd19cc399a99bdb74ac976