Support function breakpoints in edb_server_break
Summary:
# Context
The `edb_server_break` module is used to to keep track of the reasons why we have set an actual breakpoint on the VM. At the moment, we only have user-requrested line-breakpoints and breakpoints added for supporting stepping.

We want to add function breakpoints, and the way they work is that for each function clause, we will set a breapoint on the first exectuable line of each clause.

# This diff
We extend `edb_server_break` so that it can be used to set/clear functions on breakpoints and will track them. At the moment, nothing ever sets a function-breakpoint, so there is no user-facing change from this.

Reviewed By: TheGeorge

Differential Revision: D88307025

fbshipit-source-id: 9f6b74af3316fa5d67d092a6228115dd0e90706f