Better organize the edb_server_SUITE file
Summary:
# Context
The `edb_server_SUITE` currently contains all tests for function exported from `edb_server` directly (instead of the public API `edb`), which is essentially all things related to supporting mocking-like libraries

# Problem
It has a lot of testcases, but they are all just lumped in random locations, so it is hard to figured out what is being tested and what not. As shown in the next diff, there was stuff that we were not testing which let a regression slip earlier

# This diff
Group testcases thematically, using CT groups, but also inside the file. This just shuffles code around, doesn't change any behaviour

Reviewed By: michalmuskala

Differential Revision: D88558541

fbshipit-source-id: df7d6647e0722d4d5105a34d3c28bd6f254ba731