20/n][BUGFIX] Report user-module in paused event if breaking on a substitute
Summary:
# Context
When a process hits a breakpoint, an event is generated so that the debugger can find out about this. The event contains the MFA where the breakpoint happened.

To support mocking libraries, there is a notion of "module substitute". For example, when mocking `foo` with `meck`, we can declare `foo_meck` to be a "substitute" of `foo`.

# Problem
When a breakpoint is hit on a substitute module like `foo_meck`, the MFA should report to the user that it happened in `foo`, not the substitute `foo_meck`. However, this is what currently was going on.

# This diff
Add tests to ensure this doesn't regress again, and fix it so that we report the correct module.

Reviewed By: TheGeorge

Differential Revision: D88559692

fbshipit-source-id: 5ea8c386123bf2e65ed8bb00476b663c6ddf8cc4