Add DAP support for function breakpoints
Summary:
# Context
The `edb` API now supports functions breakpoints.

# Problem
This is not yet exposed via the DAP interface, so can't be used in practice

# This diff
We implement the `setFunctionBreakpoint` request and make it so that we generate an event with the `function_breakpoint` reason when such a breakpoint is hit.

We try to be flexible with the ways the user can specify the "name" of the function where to add a breakpoint, and support:

* `{M, F, A}`,
* `fun M:F/A`, and
* `M:F/A`

Reviewed By: TheGeorge

Differential Revision: D88635691

fbshipit-source-id: 68df90c8fbba02cedafec973b2e432946fd5be75