Simplify breakpoint state representation
Summary:
# Context
The `edb_server_break` module keeps track of all the breakpoints that have been set on the VM. A breakpoint could have been set for multiple reasons, like a user request, or supporting stepping on some processes.

So the most important piece of data is a (curried) map that tells us, for each module and each line with a breakpoint, what are the reasons for the breakpoint.

For stepping-breakpoints, we also need to know what we expect the stack-trace to look like before actually stopping (e.g., to avoid stopping in a recursive call when stepping-over).

# Problem
The expected call-stack patterns are currently held in a separate field, instead of being associated to the "reason" for the breakpoint.

Putting it all in one place, has a few advantages:
* We make previously invalid states unrepresentable (a step for a pid that shows up in one field but not the other)
* Once we add function-exceptions as a new reason, we will be able to encode the MFA in the metadata
* As we will have only one curried map, we can get rid of the edb_server_map module and just specialize the functions for the one structure we use for this

# This diff
Remove the `steps` field and store the patterns in the `vm_breakpoints()` structure instead.

Reviewed By: TD5

Differential Revision: D88302273

fbshipit-source-id: e5a282f69f2d393f2738871afa31ab396727a235