Add support for finding first breakable line of each function clause
Summary:
# Context
At the moment, the user is only able to set breakpoints on lines. We would like to add support for adding breakpoints on "functions", as this make it more convenient to deal with functions with lots of clauses and supports well debugging-automation workflows

# Problem
The VM exposes the lines that admit breakpoints for each function via the `erl_debugger:breakpoints/1` BIF. However, it doesn't tell us which line correspond to which clause, as the VM is unaware of "clauses"

# This diff
We can leverage the `abstract_code` section of the beam files that is added via the `+debug_info` flag. Essentially, we can find in the AST the function we are interested, from there get its clauses, and from the clause location get the first breakable line of each clause.

We will use this in the next diffs

Reviewed By: TheGeorge

Differential Revision: D88070257

fbshipit-source-id: d1ca129ed981758bc71e4d0bf77a959554ec2037